Detect when cloud servers become non-responsive and destroy them.
cloud.Server.Alive() now takes an optional bool which tries to SSH to the
server to confirm it's really alive, and Destroy()s it if not.

Destroyed servers now cause on-going RunCmd() to return an error.

New cloud.Server.SSHSession() method used by RunCmd() and Alive(), which times
out the attempt to connect with ssh.