Fix performance reversion when dealing with many jobs. Increase performance.
Remove performance sapping mutex lock in Server.reply(), replacing with proper
locks and copies of the structs being encoded.

Separate mutexes for readyaddedcallback and server state.

CLI jarchive does the decrementGroupCount in a go routine, since waiting for it
to finish is not necessary for the client, improving performance.

Fix data race in Server.logTimings().