Remove hard-coded debugging.
It was a way to attempt to save the logs at the state they were when
it crashed, should the manager pod crash & restart for any reason.
I decided to not take it any further as I couldn’t find a way to get
a pre stop trigger to run on a failure / crash. Currently, the logs
will persist across crashes, and I’ve deemed this enough.