Resolve #120: Buried jobs stay buried after a manager restart.
Job buried state is recorded to disk, so that if the manager dies and is
restarted, jobs that were buried remain buried instead of restarting.
(Jobs that were running are not yet handled properly; you still have to kill
all existing runners before restarting the manager, and those jobs start from
scratch.)

queue.Add() and AddMany() can now be told to start an item in the Run or Bury
sub-queues, for queue recovery purposes.

Memory/time/disk learning now only occurs on successful jobs, or those that
failed because they ran out of one of those resources.

Increases to job memory/time/disk due to using too much of one of those now
occur in the ready added callback along with recommendation changes, and the
time increase is fixed to be an hour longer than it took, not an hour longer
than requested. Removed job.updateRecsAfterFailure().

If a job uses more than expected memory and then fails, it is now regarded as
having failed due to memory. (Jobs are still only forcefully killed if they use
up all the machine's memory.)

TestJobqueue() tests split up in to more test functions, to isolate and help
reproduce failures.