Auto-kill lost jobs if they can be confirmed dead.
Server TTR callback now checks if a job can be confirmed dead, and kills it if
so.

Added getServer() to scheduleri and ProcessNotRunngingOnHost() to Scheduler to
enable job process checking on remote servers.

Added cloud.Server NewServer() method allowing other packages to use it to SSH
to any server.

Changed server.killJob to take a *Job to avoid wasted queue calls.